Handle guest access to launch.php
The script is used to redirect back to our platform so it shouldn't allow guest
access.

If the user is not logged in, a login redirect happens (with
require_login). If the user logs in as a guest (!), a page is shown so that the
user can click Continue and login again.